Removed unecessary service interfaces and rename serviceImpl to service
- Update all imports from rest controllers and functions
