+++ backend/test-runner/src/main/java/com/unityTest/testrunner/restImpl/CaseController.java
@@ -7,13 +7,9 @@ import com.unityTest.testrunner.models.page.TestCasePage;
 import com.unityTest.testrunner.models.response.Author;
 import com.unityTest.testrunner.models.response.TestCase;
 import com.unityTest.testrunner.restApi.CaseApi;
-import com.unityTest.testrunner.serviceImpl.CaseService;
-import com.unityTest.testrunner.serviceImpl.KeycloakService;
+import com.unityTest.testrunner.service.CaseService;
+import com.unityTest.testrunner.service.KeycloakService;
 import com.unityTest.testrunner.utils.Utils;
-import org.jboss.resteasy.client.jaxrs.ResteasyClientBuilder;
-import org.keycloak.OAuth2Constants;
-import org.keycloak.admin.client.Keycloak;
-import org.keycloak.admin.client.KeycloakBuilder;
 import org.keycloak.representations.AccessToken;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
