+++ backend/test-runner/src/main/java/com/unityTest/testrunner/restImpl/UserController.java
@@ -9,8 +9,7 @@ import com.unityTest.testrunner.models.page.SubmissionEventPage;
 import com.unityTest.testrunner.models.response.Author;
 import com.unityTest.testrunner.repository.SubmissionRepository;
 import com.unityTest.testrunner.restApi.UserApi;
-import com.unityTest.testrunner.serviceImpl.CaseService;
-import com.unityTest.testrunner.serviceImpl.KeycloakService;
+import com.unityTest.testrunner.service.CaseService;
 import com.unityTest.testrunner.utils.Utils;
 import com.unityTest.testrunner.utils.specification.AndSpecification;
 import org.keycloak.representations.AccessToken;
